libalsa-intf: Add support to handle numerical amix commands
The numerical amix commands from UCM files are not supported in
the current code unless they are the amix commands to set volume.
But there are other cases such as sample rate, which are defined
as numerical commands as well.  The change is to add the support
to cover rest of the use cases.

The change includes update to amix commands for BT SCO sample
rate from string type to numerical type.

Change-Id: I538658fb5e33ee013bf62861c016ccbe2f6c0ef9
CRs-fixed: 394165