alsa_sound: enable stereo to custom stereo mixing at dsp
During tunnel mode playback, based on a particular usr space
setting stereo output's left and right channel are averaged in
dsp, and the result is written back to both left and right
channels. So, after averaging, the left and right channel are
identical with the mono audio data. Added call to a mixer k-ctrl
that enable/disables above in kernel.

Change-Id: Id42442ccbacf4261307d2210acff16f65656a188